HMW-545 Moved the DisclaimerModal component into the Modal.tsx file.
