+++ app/client/src/components/shared/Modal.tsx
@@ -77,6 +77,32 @@ const contentStyles = (maxWidth: string) => css`
   }
 `;
 
+const disclaimerButtonStyles = css`
+  position: relative;
+  margin-bottom: 0;
+  padding: 0.125rem 0.375rem;
+  font-size: 0.625rem;
+  line-height: 1.25;
+  font-weight: normal;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  user-select: none;
+  color: ${colors.gray3};
+  background-color: ${colors.grayd};
+
+  :hover,
+  :focus {
+    background-color: ${colors.grayc};
+  }
+`;
+
+const helpIconStyles = css`
+  position: absolute;
+  top: 5px;
+  right: 5px;
+  color: #485566;
+`;
+
 const overlayStyles = css`
   background-color: ${colors.black(0.33)};
   position: fixed;
@@ -100,7 +126,7 @@ type Props = {
   triggerElm: ReactNode;
 };
 
-function Modal({
+export default function Modal({
   children,
   closeTitle,
   confirmEnabled = false,
@@ -183,4 +209,31 @@ function Modal({
   );
 }
 
-export default Modal;
+type DisclaimerProps = {
+  children: ReactNode;
+  infoIcon?: boolean;
+};
+
+export function DisclaimerModal({ children, infoIcon = false, ...props }: DisclaimerProps) {
+  return (
+    <Modal
+      closeTitle="Close disclaimer"
+      label="Disclaimer"
+      triggerElm={
+        infoIcon ? (
+          <i css={helpIconStyles} className="fas fa-question-circle"></i>
+        ) : (
+          <button
+            css={disclaimerButtonStyles}
+            // spread props so button’s styles (e.g. position) can be further set when used
+            {...props}
+          >
+            Disclaimer
+          </button>
+        )
+      }
+    >
+      {children}
+    </Modal>
+  );
+}
